♻️ Move dsn_connector.go to internal/wrap
